:truck: Rename is-height-overflowing.ts to overflow-utils.ts
